refactor: 응집도 개선
