+++ src/utils/applicationStorage.ts
@@ -37,3 +37,21 @@ export const postLocalStorage = (data: Omit<CardType, 'id'>) => {
 export const fetchLocalStorage = (key: string, initial = '') => {
   return JSON.parse(localStorage.getItem('cardList') ?? initial);
 };
+
+export const registerCardAlias = (alias: string, cardNumber: CardNumber) => {
+  const registerdCardNumber = getSerialNumber(cardNumber);
+  const cardList = fetchLocalStorage('cardList', '[]');
+  const currentCard = cardList.find(
+    (card: CardType) => getSerialNumber(card.cardNumber) === registerdCardNumber
+  );
+
+  const addedAliasCard = { alias, ...currentCard };
+
+  const restCardList = cardList.filter(
+    (card: CardType) => getSerialNumber(card.cardNumber) !== registerdCardNumber
+  );
+
+  const newCardList = [...restCardList, addedAliasCard];
+
+  localStorage.setItem('cardList', JSON.stringify(newCardList));
+};
